updated handlers file structure
