+++ handlers/urls.go
@@ -6,7 +6,6 @@ import (
 	"net/http"
 
 	"github.com/Mihai22125/URLShortenerAPI/data"
-	"github.com/gorilla/mux"
 )
 
 // Urls is a http.Handler
@@ -19,38 +18,6 @@ type keyURL struct{}
 // NewUrls creates a Urls handler
 func NewUrls(l *log.Logger) *Urls {
 	return &Urls{l}
-}
-
-// GetURL redirects shortened URL to original URL
-func (u *Urls) GetURL(rw http.ResponseWriter, r *http.Request) {
-	u.l.Println("Handle GET URL")
-
-	vars := mux.Vars(r)
-	shortURL := vars["shortURL"]
-
-	// check if given shortened url exists in database
-	urlData, err := data.GetURLByShort(shortURL)
-	if err != nil {
-		http.Error(rw, "shortened URL does not exist", http.StatusNotFound)
-		return
-	}
-	// redirect short url to coresponding web page
-	http.Redirect(rw, r, urlData.OriginalURL, http.StatusPermanentRedirect)
-}
-
-// AddURL adds a new URL struct in urlList
-func (u *Urls) AddURL(rw http.ResponseWriter, r *http.Request) {
-	u.l.Println("Handle POST URl")
-
-	newURL := r.Context().Value(keyURL{}).(data.URL)
-
-	shortURL := data.ShortURL(newURL.OriginalURL)
-	newURL.ShortURL = shortURL
-
-	data.AddURL(&newURL)
-
-	// return shortened URL for given link
-	rw.Write([]byte(newURL.ShortURL))
 }
 
 // MiddlewareValidateURL checks if given URL is valid
